Add files via upload
This is a Jupyter Notebook from the Coursera course Data Science with Python by IBM. I substituted Pandas code for the !WGET command to pull the online dataset. (I couldn't get !WGET to work on my local and Pandas was clean anyhow.) It similar to what we learned last semester but I think the code is cleaner ie. the 80/20 train/ test split etc.